requirements.txt
# ihui_public 项目依赖

# Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.0

# HTTP客户端
httpx>=0.24.0

# 数据库
sqlalchemy>=2.0.0
pymysql>=1.1.0

# 数据验证
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 工具库
python-dotenv>=1.0.0

# 日志
python-json-logger>=0.1.0

# 其他
asyncio>=3.10.0
typing-extensions>=4.0.0
